+++ test/data/test_hetero_data.py
@@ -6,15 +6,20 @@ from torch_geometric.data import HeteroData
 
 x_paper = torch.randn(10, 16)
 x_author = torch.randn(5, 32)
+x_conference = torch.randn(5, 8)
 
 idx_paper = torch.randint(x_paper.size(0), (100, ), dtype=torch.long)
 idx_author = torch.randint(x_author.size(0), (100, ), dtype=torch.long)
+idx_conference = torch.randint(x_conference.size(0), (100, ), dtype=torch.long)
 
 edge_index_paper_paper = torch.stack([idx_paper[:50], idx_paper[:50]], dim=0)
 edge_index_paper_author = torch.stack([idx_paper[:30], idx_author[:30]], dim=0)
-edge_index_author_paper = torch.stack([idx_paper[:30], idx_author[:30]], dim=0)
+edge_index_author_paper = torch.stack([idx_author[:30], idx_paper[:30]], dim=0)
+edge_index_paper_conference = torch.stack(
+    [idx_paper[:25], idx_conference[:25]], dim=0)
 
 edge_attr_paper_paper = torch.randn(edge_index_paper_paper.size(1), 8)
+edge_attr_author_paper = torch.randn(edge_index_author_paper.size(1), 8)
 
 
 def get_edge_index(num_src_nodes, num_dst_nodes, num_edges):
@@ -157,6 +162,57 @@ def test_hetero_data_rename():
     assert data['article'].x.tolist() == x_paper.tolist()
     edge_index = data['article', 'article'].edge_index
     assert edge_index.tolist() == edge_index_paper_paper.tolist()
+
+
+def test_hetero_data_subgraph():
+    data = HeteroData()
+    data.num_node_types = 3
+    data['paper'].x = x_paper
+    data['paper'].name = 'paper'
+    data['paper'].num_nodes = x_paper.size(0)
+    data['author'].x = x_author
+    data['author'].num_nodes = x_author.size(0)
+    data['conference'].x = x_conference
+    data['conference'].num_nodes = x_conference.size(0)
+    data['paper', 'paper'].edge_index = edge_index_paper_paper
+    data['paper', 'paper'].edge_attr = edge_attr_paper_paper
+    data['paper', 'paper'].name = 'cites'
+    data['author', 'paper'].edge_index = edge_index_author_paper
+    data['paper', 'author'].edge_index = edge_index_paper_author
+    data['paper', 'conference'].edge_index = edge_index_paper_conference
+
+    subset = {
+        'paper': torch.randperm(x_paper.size(0))[:4],
+        'author': torch.randperm(x_author.size(0))[:2]
+    }
+
+    out = data.subgraph(subset)
+
+    assert out.num_node_types == data.num_node_types
+    assert out.node_types == ['paper', 'author']
+
+    assert len(out['paper']) == 3
+    assert torch.allclose(out['paper'].x, data['paper'].x[subset['paper']])
+    assert out['paper'].name == 'paper'
+    assert out['paper'].num_nodes == 4
+    assert len(out['author']) == 2
+    assert torch.allclose(out['author'].x, data['author'].x[subset['author']])
+    assert out['author'].num_nodes == 2
+
+    assert out.edge_types == [
+        ('paper', 'to', 'paper'),
+        ('author', 'to', 'paper'),
+        ('paper', 'to', 'author'),
+    ]
+
+    assert len(out['paper', 'paper']) == 3
+    assert out['paper', 'paper'].edge_index is not None
+    assert out['paper', 'paper'].edge_attr is not None
+    assert out['paper', 'paper'].name == 'cites'
+    assert len(out['paper', 'author']) == 1
+    assert out['paper', 'author'].edge_index is not None
+    assert len(out['author', 'paper']) == 1
+    assert out['author', 'paper'].edge_index is not None
 
 
 def test_copy_hetero_data():